Refactor Polymarket resolve module structure
